refactored TaskCard.tsx
